% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybestep.R
\name{step_maybe}
\alias{step_maybe}
\title{Allows tuning the presence of a step.}
\usage{
step_maybe(
  recipe,
  step_fn,
  use,
  ...,
  trained = FALSE,
  skip = FALSE,
  id = rand_id("maybestep")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{step_fn}{A recipe step function that will be added to \code{recipe}.}

\item{use}{Whether to include \code{step_fn} or not into the \code{recipe}.}

\item{...}{Additional arguments passed to \code{step_fn}.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{bake.recipe()}}? While all operations are baked
when \code{\link[recipes:prep]{prep.recipe()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
Allows tuning the presence of a step.
}
\examples{
library(recipes)
df <- data.frame(x = letters[1:4], y = runif(length(letters[1:4])))
rec <- recipe(y ~ x, data = df) \%>\%
  step_maybe(step_dummy, use = TRUE, all_nominal_predictors())

rec \%>\% prep() \%>\% bake(df)

}
