% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybestep.R
\name{use_step}
\alias{use_step}
\title{A dials parameter to be used with \code{\link[=step_maybe]{step_maybe()}}.}
\usage{
use_step(values = c(FALSE, TRUE))
}
\arguments{
\item{values}{the possible values (TRUE or FALSE by default.)}
}
\description{
A dials parameter to be used with \code{\link[=step_maybe]{step_maybe()}}.
}
\seealso{
\code{\link[=step_maybe]{step_maybe()}}
}
